backend/pyproject.toml
[tool.poetry]
name = "Anythink Market Backend"
version = "0.0.0"
description = "Backend logic implementation for Anythink Market"
authors=["Anythink"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.13"
uvicorn = "^0.17.4"
fastapi = "^0.73.0"
pydantic = { version = "^1.8", extras = ["email", "dotenv"] }
passlib = { version = "^1.7", extras = ["bcrypt"] }
pyjwt = "^2.3"
databases = "^0.5.5"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
aiosql = "^3.3.1"
pypika = "^0.48.8"
alembic = "^1.7"
python-slugify = "^5.0"
Unidecode = "^1.3"
loguru = "^0.6.0"
requests = "^2.28.0"
gunicorn = "^20.1.0"
openai = "^0.25.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10"
autoflake = "^1.4"
wemake-python-styleguide = "^0.16.0"
mypy = "^0.931"
flake8-fixme = "^1.1"
pytest = "^7.0"
pytest-cov = "^3.0"
pytest-asyncio = "^0.18.0"
pytest-env = "^0.6.2"
pytest-xdist = "^2.4.0"
httpx = "^0.22.0"
asgi-lifespan = "^1.0.1"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=100
  --numprocesses=auto
  --asyncio-mode=auto
'''
env = [
  "SECRET_KEY=secret",
  "MAX_CONNECTIONS_COUNT=1",
  "MIN_CONNECTIONS_COUNT=1"
]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
